[SGH-17] Added authorization check to ProductController - delete fnc; fixed pipeline

+++ src/main/java/edu/codespring/sportgh/controller/ProductController.java
@@ -5,6 +5,7 @@ import edu.codespring.sportgh.dto.ProductOutDTO;
 import edu.codespring.sportgh.dto.ProductPageOutDTO;
 import edu.codespring.sportgh.mapper.ProductMapper;
 import edu.codespring.sportgh.model.Product;
+import edu.codespring.sportgh.security.SecurityUtil;
 import edu.codespring.sportgh.service.ProductService;
 import edu.codespring.sportgh.service.UserService;
 import jakarta.validation.Valid;
@@ -30,28 +31,29 @@ public class ProductController {
     private final ProductService productService;
     private final ProductMapper productMapper;
     private final UserService userService;
+    private final SecurityUtil securityUtil;
 
     @GetMapping
     public ResponseEntity<ProductPageOutDTO> findPageByParams(
-            @RequestParam("orderBy") Optional<String> orderBy,
-            @RequestParam("direction") Optional<String> direction,
-            @RequestParam("pageNumber") Optional<Integer> pageNumber,
-            @RequestParam("Subcategory") Optional<String[]> subcategoryNames,
-            @RequestParam("MinPrice") Optional<Double> minPrice,
-            @RequestParam("MaxPrice") Optional<Double> maxPrice,
-            @RequestParam("TextSearch") Optional<String> textSearch
+        @RequestParam("orderBy") Optional<String> orderBy,
+        @RequestParam("direction") Optional<String> direction,
+        @RequestParam("pageNumber") Optional<Integer> pageNumber,
+        @RequestParam("Subcategory") Optional<String[]> subcategoryNames,
+        @RequestParam("MinPrice") Optional<Double> minPrice,
+        @RequestParam("MaxPrice") Optional<Double> maxPrice,
+        @RequestParam("TextSearch") Optional<String> textSearch
     ) {
         return new ResponseEntity<>(
-                productService.findPageByParams(
-                        orderBy.orElse(null),
-                        direction.orElse(null),
-                        pageNumber.orElse(1),
-                        subcategoryNames.orElse(null),
-                        minPrice.orElse(null),
-                        maxPrice.orElse(null),
-                        textSearch.orElse(null)
-                ),
-                HttpStatus.OK
+            productService.findPageByParams(
+                orderBy.orElse(null),
+                direction.orElse(null),
+                pageNumber.orElse(1),
+                subcategoryNames.orElse(null),
+                minPrice.orElse(null),
+                maxPrice.orElse(null),
+                textSearch.orElse(null)
+            ),
+            HttpStatus.OK
         );
 
     }
@@ -97,12 +99,16 @@ public class ProductController {
         return new ResponseEntity<>(productMapper.productToOut(product), HttpStatus.OK);
     }
 
-    @PreAuthorize("authentication.principal.id == 404 or hasRole('ADMIN')")
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
     @DeleteMapping(path = "/{productId}")
     public ResponseEntity<?> delete(@PathVariable Long productId) {
         Product product = productService.findById(productId);
         if (product == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
+        if (!securityUtil.isCurrentlyLoggedIn(product.getUser())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
         }
 
         productService.delete(product);
